Function pallet_evm_test_vector_support::test_precompile_test_vectors

pub fn test_precompile_test_vectors<P: Precompile>(
    filepath: &str
) -> Result<(), String>

Tests a precompile against the ethereum consensus tests defined in the given file at filepath. The file is expected to be in JSON format and contain an array of test vectors, where each vector can be deserialized into an "EthConsensusTest".